refactorL routing, add marks, move tab bar controller inside tab bar coordinator, add main actor for classes
